# backend/docs/swagger.yaml
basePath: /
definitions:
  db.GetSymptomsForPatientDiseaseInstanceRow:
    properties:
      symptomDescription:
        $ref: '#/definitions/pgtype.Text'
      symptomID:
        type: integer
      symptomName:
        type: string
    type: object
  db.ListGeneralSymptomsForPatientRow:
    properties:
      createdAt:
        $ref: '#/definitions/pgtype.Timestamp'
      reportedDate:
        $ref: '#/definitions/pgtype.Date'
      symptomDescription:
        $ref: '#/definitions/pgtype.Text'
      symptomID:
        type: integer
      symptomName:
        type: string
      updatedAt:
        $ref: '#/definitions/pgtype.Timestamp'
    type: object
  db.PatientDisease:
    properties:
      createdAt:
        $ref: '#/definitions/pgtype.Timestamp'
      diagnosisDate:
        $ref: '#/definitions/pgtype.Date'
      diseaseID:
        type: integer
      notes:
        $ref: '#/definitions/pgtype.Text'
      patientDiseaseID:
        type: integer
      patientID:
        type: integer
      updatedAt:
        $ref: '#/definitions/pgtype.Timestamp'
    type: object
  db.PatientDiseaseSymptom:
    properties:
      createdAt:
        $ref: '#/definitions/pgtype.Timestamp'
      id:
        type: integer
      patientDiseaseID:
        type: integer
      symptomID:
        type: integer
      updatedAt:
        $ref: '#/definitions/pgtype.Timestamp'
    type: object
  db.PatientSymptom:
    properties:
      createdAt:
        $ref: '#/definitions/pgtype.Timestamp'
      id:
        type: integer
      patientID:
        type: integer
      reportedDate:
        $ref: '#/definitions/pgtype.Date'
      symptomID:
        type: integer
      updatedAt:
        $ref: '#/definitions/pgtype.Timestamp'
    type: object
  pgtype.Date:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
...
  server.HTTPError:
    properties:
      error:
        description: Error message
        example: Resource not found
        type: string
    type: object
  server.LinkSymptomToDiseaseInstanceRequest:
    properties:
      symptom_id:
        type: integer
    type: object
...
  server.PatientDiseaseInstanceResponse:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      diagnosis_date:
        $ref: '#/definitions/pgtype.Date'
      disease_code:
        description: Added from join
        type: string
      disease_id:
        type: integer
      disease_name:
        description: Added from join
        type: string
      notes:
        description: Keep as pointer
        type: string
      patient_disease_id:
        type: integer
      patient_id:
        description: Added for context
        type: integer
      updated_at:
        $ref: '#/definitions/pgtype.Timestamp'
    type: object
...
  server.RecordPatientDiseaseInstanceRequest:
    properties:
      diagnosis_date:
        description: Use pointer for optional date
        type: string
      disease_id:
        type: integer
      notes:
        description: Use pointer for optional notes
        type: string
    type: object
  server.RecordPatientSymptomRequest:
    properties:
      reported_date:
        description: Use pointer for optional date
        type: string
      symptom_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Service for managing patients, symptoms, and diseases.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Patient API Service
  version: "1.0"
paths:
  /disease-instances/{instanceID}:
    delete:
      consumes:
      - application/json
      description: Remove a specific patient_disease entry by its unique ID (patient_disease_id).
        This also removes associated symptom links via ON DELETE CASCADE.
      parameters:
      - description: Patient Disease Instance ID
        format: int32
        in: path
        name: instanceID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content (Successful removal)
          schema:
            type: string
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Delete a specific disease instance record
      tags:
      - Patient Relationships
  /disease-instances/{instanceID}/symptoms:
    get:
      consumes:
      - application/json
      description: Get all symptoms linked to a specific patient_disease record (diagnosis
        instance).
      parameters:
      - description: Patient Disease Instance ID
        format: int32
        in: path
        name: instanceID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved symptoms for the instance
          schema:
            items:
              $ref: '#/definitions/db.GetSymptomsForPatientDiseaseInstanceRow'
            type: array
        "400":
          description: Invalid Instance ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Disease instance not found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: List symptoms for a specific disease instance
      tags:
      - Patient Relationships
    post:
      consumes:
      - application/json
      description: Create an association between an existing symptom and a specific
        patient disease instance.
      parameters:
      - description: Patient Disease Instance ID
        format: int32
        in: path
        name: instanceID
        required: true
        type: integer
      - description: Symptom ID to link
        in: body
        name: link
        required: true
        schema:
          $ref: '#/definitions/server.LinkSymptomToDiseaseInstanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Symptom linked successfully
          schema:
            $ref: '#/definitions/db.PatientDiseaseSymptom'
        "400":
          description: Invalid Instance ID or request payload
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Disease instance or Symptom not found (FK constraint)
          schema:
            $ref: '#/definitions/server.HTTPError'
        "409":
          description: Symptom already linked to this instance (unique constraint)
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Link a symptom to a disease instance
      tags:
      - Patient Relationships
  /disease-instances/{instanceID}/symptoms/{symptomID}:
    delete:
      consumes:
      - application/json
      description: Remove the association between a symptom and a specific patient
        disease instance.
      parameters:
      - description: Patient Disease Instance ID
        format: int32
        in: path
        name: instanceID
        required: true
        type: integer
      - description: Symptom ID to unlink
        format: int32
        in: path
        name: symptomID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content (Successful removal)
          schema:
            type: string
        "400":
          description: Invalid Instance ID or Symptom ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Unlink a symptom from a disease instance
      tags:
      - Patient Relationships
...
  /patient-symptoms/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a specific patient_symptoms entry by its unique ID.
      parameters:
      - description: Patient Symptom Record ID
        format: int32
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content (Successful removal)
          schema:
            type: string
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Delete a specific general symptom record
      tags:
      - Patient Relationships
...
  /patients/{patientID}/disease-instances:
    get:
      consumes:
      - application/json
      description: Get all recorded disease instances (diagnoses) for a given patient
        ID.
      parameters:
      - description: Patient ID
        format: int32
        in: path
        name: patientID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved disease instances
          schema:
            items:
              $ref: '#/definitions/server.PatientDiseaseInstanceResponse'
            type: array
        "400":
          description: Invalid Patient ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: List disease instances for a specific patient
      tags:
      - Patient Relationships
    post:
      consumes:
      - application/json
      description: Create a record of a specific disease diagnosis for a patient,
        with optional date and notes.
      parameters:
      - description: Patient ID
        format: int32
        in: path
        name: patientID
        required: true
        type: integer
      - description: Disease ID, optional diagnosis date and notes
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/server.RecordPatientDiseaseInstanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Disease instance recorded successfully
          schema:
            $ref: '#/definitions/db.PatientDisease'
        "400":
          description: Invalid Patient ID or request payload
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Patient or Disease not found (FK constraint)
          schema:
            $ref: '#/definitions/server.HTTPError'
        "409":
          description: Duplicate instance for this patient/disease/date (unique constraint)
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Record a disease instance for a patient
      tags:
      - Patient Relationships
  /patients/{patientID}/general-symptoms:
    get:
      consumes:
      - application/json
      description: Get all symptoms associated with a given patient ID via the patient_symptoms
        table (not linked to specific diagnoses).
      parameters:
      - description: Patient ID
        format: int32
        in: path
        name: patientID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved general symptoms for patient
          schema:
            items:
              $ref: '#/definitions/db.ListGeneralSymptomsForPatientRow'
            type: array
        "400":
          description: Invalid Patient ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: List general symptoms reported by a patient
      tags:
      - Patient Relationships
    post:
      consumes:
      - application/json
      description: Associate an existing symptom with a patient (not tied to a specific
        diagnosis instance). Optionally include the date it was reported.
      parameters:
      - description: Patient ID
        format: int32
        in: path
        name: patientID
        required: true
        type: integer
      - description: Symptom ID and optional reported date
        in: body
        name: symptom
        required: true
        schema:
          $ref: '#/definitions/server.RecordPatientSymptomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Symptom recorded successfully
          schema:
            $ref: '#/definitions/db.PatientSymptom'
        "400":
          description: Invalid Patient ID or request payload
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Patient or Symptom not found (FK constraint)
          schema:
            $ref: '#/definitions/server.HTTPError'
        "409":
          description: Relationship already exists for this date (unique constraint)
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Record a general symptom for a patient
      tags:
      - Patient Relationships
...
schemes:
- http
- https
swagger: "2.0"